Add support for Django 1.9+ which removed ValuesQuerySet.
