+++ bananas/utils.py
@@ -1,0 +1,99 @@
+from itertools import chain
+
+MISSING = object()
+
+
+class ModelDict(dict):
+
+    _nested = None
+
+    def __getattr__(self, item):
+        """
+        Try to to get attribute as key item.
+        Fallback on prefixed nested keys.
+        Finally fallback on real attribute lookup.
+        """
+        try:
+            return self.__getitem__(item)
+        except KeyError:
+            try:
+                return self.__getnested__(item)
+            except KeyError:
+                return self.__getattribute__(item)
+
+    def __getnested__(self, item):
+        """
+        Find existing items prefixed with given item
+        and return a new ModelDict containing matched keys,
+        stripped from prefix.
+
+        :param str item: Item prefix key to find
+        :return ModelDict:
+        """
+        # Ensure _nested cache
+        if self._nested is None:
+            self._nested = {}
+
+        # Try to get previously accessed/cached nested item
+        value = self._nested.get(item, MISSING)
+
+        if value is not MISSING:
+            # Return previously accessed nested item
+            return value
+
+        else:
+            # Find any keys matching nested prefix
+            prefix = item + '__'
+            keys = [key for key in self.keys() if key.startswith(prefix)]
+
+            if keys:
+                # Construct nested dict of matched keys, stripped from prefix
+                n = ModelDict({key[len(item) + 2:]: self[key] for key in keys})
+
+                # Cache and return
+                self._nested[item] = n
+                return n
+
+        # Item not a nested key, raise
+        raise KeyError(item)
+
+    @classmethod
+    def from_model(cls, model, *fields, **named_fields):
+        """
+        Work-in-progress constructor,
+        consuming fields and values from django model instance.
+        """
+        d = ModelDict()
+
+        if not (fields or named_fields):
+            # Default to all fields
+            fields = [f.attname for f in model._meta.concrete_fields]
+
+        not_found = object()
+
+        for name, field in chain(zip(fields, fields), named_fields.items()):
+            _fields = field.split('__')
+            value = model
+            for i, _field in enumerate(_fields, start=1):
+                # NOTE: we don't want to rely on hasattr here
+                previous_value = value
+                value = getattr(previous_value, _field, not_found)
+
+                if value is not_found:
+                    if _field in dir(previous_value):
+                        raise ValueError(
+                            '{!r}.{} had an AttributeError exception'
+                            .format(previous_value, _field))
+                    else:
+                        raise AttributeError(
+                            '{!r} does not have {!r} attribute'
+                            .format(previous_value, _field))
+
+                elif value is None:
+                    if name not in named_fields:
+                        name = '__'.join(_fields[:i])
+                    break
+
+            d[name] = value
+
+        return d
